requirement.txt
aiohttp
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1

certifi==2021.10.8
charset-normalizer==2.0.12
cytoolz
Deprecated==1.2.13





django-ipware==4.0.2
django-redis==5.2.0
python-dotenv==0.21.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.2
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
protobuf==3.19.4

pycryptodome==3.14.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
redis==4.1.4

requests==2.27.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
typing-extensions==4.1.1
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
whitenoise==6.0.0
wrapt==1.13.3
yarl
zipp==3.7.0
dj-database-url==1.2.0
